Renamed SettingsEditor to SettingsScreen
